Extract gdns file generation to separate class
