+++ tools/entry-generator/src/main/kotlin/org/godotengine/kotlin/entrygenerator/EntryGenerator.kt
@@ -1,12 +1,9 @@
 package org.godotengine.kotlin.entrygenerator
 
 import com.squareup.kotlinpoet.FileSpec
-import de.jensklingenberg.mpapt.common.findAnnotation
 import de.jensklingenberg.mpapt.model.Element
-import org.godotengine.kotlin.annotation.RegisterClass
 import org.godotengine.kotlin.entrygenerator.generator.*
 import org.jetbrains.kotlin.name.Name
-import org.jetbrains.kotlin.resolve.descriptorUtil.fqNameSafe
 import java.io.File
 
 
@@ -44,7 +41,7 @@ class EntryGenerator {
 
         properties.forEachIndexed { index, propertyElement ->
             val visibleInEditor: Boolean = propertyElement
-                    .annotation!!
+                    .annotation!! //we already know here that the annotation is present. Otherwise this property would not have been included in this set
                     .allValueArguments
                     .getValue(Name.identifier("visibleInEditor"))
                     .value as Boolean
@@ -59,39 +56,10 @@ class EntryGenerator {
 
         gdNativeFunctionBindingGenerator.generateGDNativeBindingFunctions(entryFileSpec)
 
-        generateGdnsFiles(godotProjectPath, libraryPath, classes)
+        GdnsFileGenerator().generateGdnsFiles(godotProjectPath, libraryPath, classes)
 
         entryFileSpec
                 .build()
                 .writeTo(File(kaptGeneratedDirectory))
-    }
-
-    private fun generateGdnsFiles(outputDirectory: String, libraryPath: String, classes: Set<Element.ClassElement>) {
-        classes.forEach { classElement ->
-            val pathFromAnnotation = classElement
-                    .classDescriptor
-                    .annotations
-                    .findAnnotation(RegisterClass::class.java.name)!!
-                    .allValueArguments
-                    .filter { it.key.asString() == "outputDir" }
-                    .map { it.value.value as String }
-                    .firstOrNull { it.isNotBlank() }
-
-            val outputPath = pathFromAnnotation ?: "scripts/${(classElement.classDescriptor.containingDeclaration.name.asString()).replace(".", File.separator)}"
-
-            val out = File("$outputDirectory/$outputPath/${classElement.classDescriptor.name}.gdns")
-            out.parentFile.mkdirs()
-            out.createNewFile()
-            out.writeText("""
-                [gd_resource type="NativeScript" load_steps=2 format=2]
-                
-                [ext_resource path="res://$libraryPath" type="GDNativeLibrary" id=1]
-                
-                [resource]
-                resource_name = "${classElement.classDescriptor.name}"
-                class_name = "${classElement.classDescriptor.fqNameSafe}"
-                library = ExtResource( 1 )
-            """.trimIndent())
-        }
     }
 }
